refactor(cli,app-services): move oversized inline test modules into sibling files
`serve.rs` (315 LOC) and `launch_options.rs` (327 LOC) both sit above the
300 LOC ratchet in `scripts/rust-complexity-baseline.txt`, where the gate
fails on any growth at all. Both files are about to take new code — a
default inference profile and the sampling-override fix — so they need
headroom before that work can land, not after.

The sibling `*_tests.rs` split is the convention this repo already uses
seven times over (`sampling_params`, `reasoning_args`, `profiles`,
`update`, `explain_display`, `render`, `render_reasoning`), for exactly
this reason: the test block outgrew the source and was making the source
itself hard to read.

Pure move. The 17 tests are byte-identical apart from one level of
indentation, and `cargo test` reports the same names and the same count
before and after — 10 under `serve_tests`, 7 under `launch_options_tests`.
`serve.rs` drops to 147 and `launch_options.rs` to 141, which takes both
out of the ratchet's scan entirely.

Full pre-commit green.
